@@ -1,5 +1,4 @@
 namespace EVIL.Grammar.AST.Base;
 
-public abstract class ConstantExpression : Expression
-{
-}
+public abstract class ConstantExpression 
+    : Expression;
