@@ -2,12 +2,8 @@ namespace EVIL.Grammar.AST.Constants;
 
 using EVIL.Grammar.AST.Base;
 
-public sealed class BooleanConstant : ConstantExpression
+public sealed class BooleanConstant(bool value) 
+    : ConstantExpression
 {
-    public bool Value { get; }
-
-    public BooleanConstant(bool value)
-    {
-        Value = value;
-    }
+    public bool Value { get; } = value;
 }
