@@ -2,6 +2,4 @@ namespace EVIL.Grammar.AST.Constants;
 
 using EVIL.Grammar.AST.Base;
 
-public sealed class NilConstant : ConstantExpression
-{
-}
+public sealed class NilConstant : ConstantExpression;
