@@ -3,12 +3,8 @@ namespace EVIL.Grammar.AST.Constants;
 using EVIL.CommonTypes.TypeSystem;
 using EVIL.Grammar.AST.Base;
 
-public class TypeCodeConstant : ConstantExpression
+public class TypeCodeConstant(DynamicValueType value) 
+    : ConstantExpression
 {
-    public DynamicValueType Value { get; }
-
-    public TypeCodeConstant(DynamicValueType value)
-    {
-        Value = value;
-    }
+    public DynamicValueType Value { get; } = value;
 }
