@@ -2,13 +2,11 @@ namespace EVIL.evil;
 
 using System;
 
-public class RuntimeModuleLoadException : Exception
+public class RuntimeModuleLoadException(
+    string message,
+    Exception innerException,
+    string filePath
+) : Exception(message, innerException)
 {
-    public string FilePath { get; }
-
-    public RuntimeModuleLoadException(string message, Exception innerException, string filePath) 
-        : base(message, innerException)
-    {
-        FilePath = filePath;
-    }
+    public string FilePath { get; } = filePath;
 }
