@@ -5,7 +5,7 @@ using EVIL.Grammar.AST.Expressions;
 
 public partial class Compiler
 {
-    public override void Visit(ConditionalExpression conditionalExpression)
+    protected override void Visit(ConditionalExpression conditionalExpression)
     {
         var elseLabel = Chunk.CreateLabel();
         var endLabel = Chunk.CreateLabel();
