@@ -5,7 +5,7 @@ using EVIL.Grammar.AST.Expressions;
 
 public partial class Compiler
 {
-    public override void Visit(ExtraArgumentsExpression extraArgumentsExpression)
+    protected override void Visit(ExtraArgumentsExpression extraArgumentsExpression)
     {
         var mode = extraArgumentsExpression.UnwrapOnStack ? (byte)1 : (byte)0;
         Chunk.CodeGenerator.Emit(OpCode.XARGS, mode);
