@@ -4,7 +4,7 @@ using EVIL.Grammar.AST.Expressions;
 
 public partial class Compiler
 {
-    public override void Visit(SymbolReferenceExpression symbolReferenceExpression)
+    protected override void Visit(SymbolReferenceExpression symbolReferenceExpression)
     {
         EmitVarGet(symbolReferenceExpression.Identifier);
     }
