@@ -5,7 +5,7 @@ using EVIL.Grammar.AST.Expressions;
 
 public partial class Compiler
 {
-    public override void Visit(TableExpression tableExpression)
+    protected override void Visit(TableExpression tableExpression)
     {
         Chunk.CodeGenerator.Emit(OpCode.TABNEW);
 
