@@ -5,7 +5,7 @@ using EVIL.Grammar.AST.Expressions;
 
 public partial class Compiler
 {
-    public override void Visit(TypeOfExpression typeOfExpression)
+    protected override void Visit(TypeOfExpression typeOfExpression)
     {
         Visit(typeOfExpression.Target);
 
