@@ -4,7 +4,7 @@ using EVIL.Grammar.AST.Miscellaneous;
 
 public partial class Compiler
 {
-    public override void Visit(ArgumentList argumentList)
+    protected override void Visit(ArgumentList argumentList)
     {
         for (var i = 0; i < argumentList.Arguments.Count; i++)
         {
