@@ -4,7 +4,7 @@ using EVIL.Grammar.AST.Statements.TopLevel;
 
 public partial class Compiler
 {
-    public override void Visit(AttributeList attributeList)
+    protected override void Visit(AttributeList attributeList)
     {
         foreach (var attr in attributeList.Attributes)
             Visit(attr);
