@@ -5,7 +5,7 @@ using EVIL.Grammar.AST.Statements;
 
 public partial class Compiler
 {
-    public override void Visit(SkipStatement skipStatement)
+    protected override void Visit(SkipStatement skipStatement)
     {
         Chunk.CodeGenerator.Emit(OpCode.JUMP, Loop.StartLabel);
     }
