@@ -5,7 +5,7 @@ using EVIL.Grammar.AST.Statements;
 
 public partial class Compiler
 {
-    public override void Visit(ThrowStatement throwStatement)
+    protected override void Visit(ThrowStatement throwStatement)
     {
         Chunk.MarkThrowing();
             
